move api dir to auth
